theme/requirements.txt
chess
Pillow
